backend/pyproject.toml
[project]
name = "genreader-backend"
version = "0.1.0"
description = "GenReader backend: VLM-powered OCR with coordinate localization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "arq>=0.26",
    "redis>=5.2",
    "minio>=7.2",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.28",
    "pillow>=11.0",
    "pymupdf>=1.25",
    "numpy>=2.1",
    "slowapi>=0.1.9",
    "websockets>=13.1",
]

[project.optional-dependencies]
ml = [
    "torch>=2.5",
    "transformers>=4.46",
    "paddleocr>=2.9",
    "paddlepaddle>=2.6",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
